+++ pilz_industrial_motion_testutils/include/pilz_industrial_motion_testutils/gripper.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018 Pilz GmbH & Co. KG
+ * Copyright (c) 2019 Pilz GmbH & Co. KG
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,22 +14,20 @@
  * limitations under the License.
  */
 
-#ifndef DEFAULT_VALUES_H
-#define DEFAULT_VALUES_H
+#ifndef GRIPPER_H
+#define GRIPPER_H
 
-/*
- * @brief This file contains all default values needed for testing.
- */
+#include "ptp.h"
+#include "jointconfiguration.h"
+
 namespace pilz_industrial_motion_testutils
 {
 
-static constexpr double DEFAULT_VEL {0.01};
-static constexpr double DEFAULT_ACC {0.01};
-static constexpr double DEFAULT_BLEND_RADIUS {0.01};
+class Gripper : public Ptp<JointConfiguration, JointConfiguration>
+{
 
-static constexpr double DEFAULT_VEL_GRIPPER {0.5};
-static constexpr double DEFAULT_ACC_GRIPPER {0.8};
+};
 
 }
 
-#endif // DEFAULT_VALUES_H
+#endif // GRIPPER_H